style(settings): remove tint from drawables (#1020)
colorControlNormal attribute could not be resolved and was causing build failures when selecting "Compile all sources" or "gradlew build".  colorControlNormal is found in the Android theme, but the settings submodule did not have any theme, so it could not be resolved.

The tint is applied by Jetpack Compose Icon, so it's not needed in the drawable XML.

Co-authored-by: rcrosbie <1092cameo@gmail.com>